Add the concept of sublanguages
Sublanguages can be used in addition to the configured language of the
driver (e.g. Fortran + CUDA or C + OpenCL).

Sublanguages may have their own profile otherwise they will use the
configured one.

OpenCL and CUDA are the only sublanguages defined now. OpenCL does
not have a profile but is already not handled at all, so this is
not a big deal.

CUDA uses the profile 'cuda' which simply invokes nvcc. Extra flags
to that profile can be passed as usual using --Wx:cuda:

See #1594