Do not expand array constants
Although we have to compute the constant values (for proper constant
calculation in the frontend) we will retain the original tree and
properly keep the constant.

This means that some nodes may have unconventional constant values
attached to them:

   INTEGER :: A(10) = 1

Will have a NODECL_INTEGER_LITERAL as initialization tree but the
constant attached to it won't be just an integer but an a array of
integers.

On the other hand,

  INTEGER :: A(10) = (/ (I, I=1,10) /)

Will have a NODECL_STRUCTURED_VALUE with an item inside of kind
NODECL_FORTRAN_IMPLIED_DO. The constant value attached to the
structured-value will be an array as well.

This commit fixes #1654