Fixes and improvements related to new 57.config.omp.smp configuration file
Commit [36abbaaad56649b6c18211ba8aa181835bc7bfa6] changes
introduces 57.config.omp.smp as the replacement of
57.config.omp.all. The file was missing from configure,
this commit fixes this.

This commit also adds an extra check when installing configurations
so the old 57.config.omp.all is removed, lest it was already
in the install dir from a previous installation.

See #1678