Modify ExtensibleGraph internal for robustness
Some methods inside ExtensibleGraph class perform traversals in the PCFG.
For that, they need an attribute in Node indicating whether a given node
has been visited or not. In order to avoid interferences between other
analyses' traversals and calls to the ExtensibleGraph methods that
perform traversals, we have created two new attributes in the Node class
that are intended to be used only by ExtensibleGraph.