Implement task concurrency analysis for liveness
Liveness common equations may vary when OpenMP tasks, for the
order of execution in the PCFG is not stablished for the edges
that connect a task with the rest of the graph.
Because of that, the mechanism that calculates concurrent tasks
that previously was used during auto-scoping analysis, must come
early, before liveness is computed.
Thus, liveness analysis calls first this mechanism, creating
a map in the ExtensibleGraph with the tuple:
<Node* task, ObjectList<Node*> concurrent_tasks>.

The equations that vary are the ones computing the LiveIn variables
of a node 'x' when:
- 'x' is in a task or is a task:
  LI(x) = UE(x) U ( LO(x) - KILL(x) - ( U KILL(y) ) ),
          where 'y' = all concurrent task of 'x'
- 'x' is a task creation:
  LI(x) = UE(x) U ( LO(x) - KILL(x) - KILL(y) ),
          where 'y' = task created by 'x'