Fix UseDef analysis for TaskCreation nodes
Due to the introduction of TaskCreation nodes in the PCFG,
the UseDef analysis had become incorrect. So far, the usage was
computed taking into account the information in a given node.
For easyness, we propagate the usage information in a task to
its task creation node. And we propagate information bottom-top
in concatenated task_creation nodes.