Create __MERCURIUM_PRETTY_FUNCTION__
__PRETTY_FUNCTION__ is a predetermined identifier in C++ functions of
g++ which contains an adorned name of the function. Sadly such name
is very compiler specific and varies among compilers. Until know we
emitted it as a const char* but it is actually a const char [x], where
'x' is what actually varies between compilers.

As a solution, introduce a predetermined identifier called
__MERCURIUM_PRETTY_FUNCTION__ and an alias to it __PRETTY_FUNCTION__
with a value similar (in usefulness) to the one C++ compilers emit.

Note that for dependent functions __PRETTY_FUNCTION__ is left as is, as its
type and value are dependent.

This commit fixes #1734