Reduction symbol can also be Induction Variables
It is not clear how to define induction variables in parallel programs.
We have decided that a variable which is an induction variable in
sequential analysis is also an induction variable when we analyze
parallel codes. This affects also reduction variables.

Because of that, we include a new method in the analysis API to ask
about induction variables which are not reduction symbols.

See #1643