Several fixes for Use-Def analysis
1. Fixes the computation of References and ArraySubscripts usage
    * the base of an array subscript is not used unless its type is
      pointer.
    * referenced objects are not used unless they are array
      subscripts and the type of the base is a pointer.

2. Fixes usage propagation inside split nodes in the PCFG: when a
function call causes a split node, we don't need to calculate its usage
two times, we just propagate the usage computed in the function call
node to the other nodes that contain this call.

3. Enhances subobjects usage computation: when an object and some of
its subobjects (aplicable to arrays and aggregates) are inserted
to the same usage list, the subobject is removed, since the object
already contains it.

This commit also adds support in the PCFG for ObjectInit nodecls
that can produce a split node in the initialization expression.

Fixes #1748