Do not lose track of alive tasks from different domains
When an alive task of a different domain does not synchronize,
we were ignoring it, thus forgetting to add it to the alive-out-tasks.

#pragma omp task
{
  #pragma omp task
  {
    // B
  }
}
#pragma omp task // X
{
  // C
}

When computing synchronization between B and C at the task creation point
X, B was being discarded. The correct thing to do is add it as a
non-synchronized task (i.e. still alive).

This commit fixes #1781