Embed barrier and its implicit flush in graph node
OpenMP defines an implicit flush during a barrier.
In the PCFG this is represented with a flush before the barrier and
a flush after the barrier. In order to keep this nodes as a unique
transaction we embed them inside a barrier graph node.

Tasks that are synchronized in a given barrier, point to the graph
node instead of pointing to the barrier single node, as did before.