Load Nanos++ headers _after_ having parsed Fortran
This is a radical change but meaningful as it does not make sense to
pollute Fortran scope before actually needing to.

Before this commit, we loaded nanos++ headers and then parsed user
Fortran code. This approach is risky as we may end looking up global
names (like math.h functions y1) where the implicit mechanism of Fortran
should be used instead.

Now we load nanos++ headers in the run of the phase just before anything,
so we can still emit calls to Nanos++ when lowering OpenMP.

This commit fixes #1815