AutoScope phase must run analysis in all structures
Analysis was called only for tasks. This caused tasks inside other
control structures as loop or ifelse not to be analyzed.

This commit also removes unused code from the Analysis Test Phase
that was there because in the past the auto-scope analysis was called
as a service instead of as a phase, as it is now.