Change UseDef methodology
UseDef analysis is now implemented in such an order that a PCFG is
analyzed only if all functions called in the PCFG are already analyzed
or, if they are not, it is because its code is not reachable or
it is a recursive call.

This commit also shortens the line length for the analysis information
when it is shown in DOT format.