Mark an implicit int type
In C99, declarations may lack a type. The compiler completes the tree
adding an int specifier but this happens as a side effect, so we may end
seeing an int specifier when we expected none.

Since this scenario is indistinguishable from a true int specifier,
mark the implicit int specifier using another node.

This commit fixes #1847