Flatten sequence types when updating them
Sometimes sequences of types include non-pack types and pack types that
result in nested sequences when they are updated.

We do not allow nested sequences (although we allow a sequence having
a pack type which is the culprit of the nested expansion) so
when updating, let the constructor of sequences flatten them for us.