Fix the criteria for parenthesizing assignments in codegen
The lhs of an assignment is a logical-or while the rhs of an assignment
must be another assignment. Since codegen already does special code for
the assignments themselves, we will use this fact instead of simply
comparing the priority, which leads to odd behaviour (this was the reason
why throw and conditional expressions where defined to have the same
priority of an assignment).

See #1889