Changes in the tokenization of new[], delete[] and noexcept
If noexcept is not followed by a ( it will be TOKEN_NOEXCEPT_ALONE

This fixes the following stupid case

  template <typename T>
  void foo(T) noexcept (T::Value)

as the parser allows the following interpretation

  template <typename T>
  void (foo(T) noexcept) (T::F);

which is very wrong.

To fix this we also observed some %dprec added because of new[] and
delete[] being interpreted wrongly as well.

  void operator new[](int size_t);

could be interpreted as

  void (operator new)[](int size_t);

which is also very wrong.

Now a new followed by [] will always be TOKEN_NEW_ARRAY. Likewise for
delete[].

Also note that liberal use of whitespaces like in new [  ] or
delete [
] is also taken into account.