Fix method in ExtensibleGraph looking for nodecls
Method ExtensibleGraph::find_nodecl was using equal_nodecl to compare
nodecls. This was not totally correct because we were always comparing
at a statment level, thus missing innex expressions of the statement.
Now we are calling the function stmtexpr_contains_nodecl, that
performs the comparison not only at a statement level, but also at a
expression level.