Fix ArrayAccessInfoVisitor for nested loops
When checking the adjacency of an access, we must take into account
the point of view, which is important for the vectorization process.
This means, for example:
{{{#!c++
for(i=0; i<N; ++i)
   for(j=0; j<M; ++j)
      A[i+j]
}}}
If vectorizing the external loop, the acces A[i+j] is adjacent
because i is fixed and j increments with a step=1.
On the other hand, if vectorizing the internal loop, the access A[i+j] is
not adjacent, but constant, for both variables are fixed within the
inner loop.

Further work may be needed. Revision required.
See #1861.