Implement more accurate IV dependent access analysis
Induction Variable dependent accesses can be of two types:
1.- IV is used as modifier
2.- IV is used as condition of the modification
This commit extends the method 'is_induction_variable_dependent_access'
to take into account this two conditions.

For example:
{{{#!c++
for(i=0; ....)
   for(j=i; ...)
   {
      if(i%2)
         k=0;
      else
         k=1;
      A[j+k]
   }
}}}
In the code above, the IV from which we vectorize is 'i'.
In the access 'A[j+k]' the two variables 'j' and 'k' depend on 'i',
but in different ways:
1.- 'j' is modified with the value of 'i'
2.- 'k' is modified depending on a condition that depends on 'i'

See #1861