Make IV dependent array access analysis recursive
This commit refines the process of determining whether an
ArraySubscript depends on the value of an induction variable.
Now, this process is recursive, meaning that we check recursively
the variables used to define the variables involved in the
subscript until we reach the scope from which we are analyzing or
we finf that the access is indeed dependent of an IV.

For example, in the following code we can determine that the array
subscript 'A[M]' depends on 'i', for 'M' depends on 'l',
'l' depends on 'j' and 'j' depends on 'i':
{{{#!c++
int i, j, l, M;
for (i=0; i<N; i++) {
   for (j=i; j<K; j++) {
      l = j;
      if(l%2 == 0)
         M = 5;
      else
         M = 2;
      A[i] = A[M];
   }
}
}}}

See #1861.