Add support for simd_reduction clause
Add suport for a new kind of reductions and its clause.
I have not changed the implementation of the common reductions.
Simd reductions are maintained in a separated data structure.

In any case, this commit should be carefully reviewed by Roger.
After his approval, if he does, this commit MUST BE CHERRY-PICKED to the Master branch.
Do not close it without doing the cherry-pick!

It fixes #1936