Add support for reductions in the parallel construct
Initial support for reductions in the parallel
construct of Intel RTL. This commit must be carefully
reviewed. I introduced some warnings to highlight some
"points of interest".

Currently, the final code does not compile since the
master_combiner is not well generated. So far, we
generate something like this:

{{{
    result += p_result6;
}}}

but the 'result' variable needs a dereference:

{{{
    (*result) += p_result6;
}}}

See #1939