Fix wrong identifier in TDG JSON edge variables
The identifier of a variable referenced in a condition must correspond
to the identifier of the same variable defined in the defVars rule.

This commit also fixes the name of the function, which was the name
of the associated PCFG and now is the name of the function, if exists,
or the empty string, otherwise.