Analysis detects OmpSs reductions
New method implemented in AnalysisStaticInfo
to detect reductions with the following form:
* r += foo();
* r = foo(r);
* r = foo() + r;
where 'foo' is always an outlined task.
Unfortunately, the current implementation also recognizes
as reduction something like:
r = foo() + a[r];

To use this method it is enough to do:
   AnalysisStaticInfo a;
   bool is_reduc = a.is_ompss_reduction( n );
Although, using an AnalysisStaticInfo created with the parametrized
contructor is also valid (but not necessary).

See #1873.