OmpSs reductions analysis needs FunctionTaskSet
The analysis method meant to check whether an expression is an
OmpSs reduction or not is used before Core transforms the tree
creating TaskCall nodes. Thus, it requires a new parameter
containing all the functions in the code that are defined as tasks.

The new implementation allows determining that the use of the
supposedly reduced variable as a subscript in the rhs of the expression
does not count as a reduction:
x = f() + a[x] -> this is not a reduction

Fixes #1873