Represent the referenced subobject in NODECL_CLASS_MEMBER_ACCESS
In C++, a class-member access may become a set of
NODECL_CLASS_MEMBER_ACCESS when base classes are involved. This way
we can tell which subobject is actually being accessed.

The class-member-access-form child has been removed and instead
we have a member-literal which represents the exact C++ name used
to access the object. This child is only useful for C++ codegen.

This commit fixes #1935