Store usage of variables based on the OpenMP scope
Variables that are defined as private or firstprivate in an OpenMP
construct must be treated accordingly during Use-Def analysis.
Thus, we must not propagate to outer nodes variables which are used,
defined or have an undefined behaviour if those variables are private.
On the other hand, we must not propagate variables that are defined
or have an undefined behaviour if those variables are firstprivate.