Fix UseDef analysis for subobjects
When computing the UseDef analysis for a node with several statements
or propagating the usage of inner nodes to their outer nodes,
the analysis must take into account objects that can have subobjects such as:
- ArraySubscripts
- ClassMemberAccess
For example, we are able to handle when an array A has a part of it that
is defined(i.e. A[1]) and a part that is used (i.e. A[0:3]).

We have extended the setters of the use-definition attributes associated
to a node in such a way that we are able either to assign a new set of
variables or just add set of variables.