PCFG adds graph nodes for each Case in a Switch
It is convinient to wrap all statements within a Switch Case inside
a graph node, so, when we want to get the condition of that case
it is not necessary to backwards traverse the PCFG, but to traverse
the outer nodes of a given node.

Also for convinience, we directly link the condition nodes of a
switch statement, an ifelse statement and any kind of loop to their
graph nodes. Thus, to access the condition we do not have to traverse
the graph, but to access the attribute 'condition' of the graph node.