Add new Vector nodes to Analysis
This commit adds a set of vector nodes to PCFG and UseDef analyses.
It fixes #1987.

The following nodes are not added to PCFG because they are
clause arguments that cannot contain expressions:
* NODECL_RELAXED_FLAG
* NODECL_EVICT_FLAG
The following nodes are not added to PCFG because
they were already supported:
* NODECL_VECTOR_REDUCTION_ADD
* NODECL_VECTOR_REDUCTION_MINUS
* NODECL_VECTOR_REDUCTION_MUL

UseDef analysis only adds the following nodes, because they are
the only ones that modifify the memory:
* NODECL_VECTOR_STREAM_STRORE
* NODECL_VECTOR_SINCOS