Add another list containing the members in declaration order
When the definition of a previously declared member class (but this
would apply to forward enums of C++2011) was found, the original symbol
was removed from the member list and added at the end. This breaks
declaration order, so to counter-act this codegen forward declared
all the member classes of a class.

Unfortunately this had the side effect that we were altering the order
of instantiations inside the class, totally defeating the purpose of a
forwarded declaration. So, in a class like the following

template <typename T>
struct A
{
  struct B;
  B f();
  struct B { /* .. */ };
};

the member list ended being [f, B] rather than [B, f]. Instantiation
actually needs the list [f, B] while codegen needs something like
[B:decl, f:decl, B:def].

We fix this by avoiding juggling with the member classes, so the member
list will end being [B, f] (the repeated declaration of the member class
B does not have to appear in the member list as it is not what we want,
see an example below where this makes sense) and we add an extra list
(which actually is only meaningful for member classes and enums) that
remembers whether the declaration is actually a definition or not. For the
example above it is the already mentioned [B:decl, f:decl, B:def] list.

This new list will be used by codegen and the functions for adding members
already update it, so it should be transparent to the user except for
an extra parameter stating whether the declaration is a definition or not.
While this might be used for all member kinds it is only meaningful at
the time for classes and enums.

Regarding while a forward declaration must be preserved in the member list
the following valid example shows why.

template <typename T>
struct A
{
    struct Test;
    typedef typename Test::Moo Type;
    struct Test
    {
        typedef int Moo;
    };
};

typedef A<int>::Type T;

When instantiating A<int>, the class A<int>::Test should already
be complete as well at the point of instantiation of A<int>::Type. This
happens automatically if the member list is effectively [Test, Type].

This commit fixes #2002