Add support for char16_t and char32_t literals
This includes support both for string and character literals.

Note that currently only char and utf-8 characters are stored
as const_values of characters. All other encodings are represented
as UTF-32.

The idea is that in a future const_value_t will assume that characters are
elements of ISO/IEC 10646 (i.e. Unicode codepoints). What is unclear is what
would be the most suited representation in memory. UTF-32 is simpler regarding
the handling of codepoints but UTF-8 may be more compact.

Note also that char16_t string literals require a conversion
from UTF-32 to UTF-16LE/BE in order to compute the number of char16_t
in a string literal. We do this conversion using 'iconv'. Such conversion
is thrown away.