First test case working implementation of RangeAnalysis
This commit adds information about constraints determining the
possible values of a variable at any given point of a program.
At any node of the PCFG a set of nodecls and one only constraint
associated may be attached. Note that this constraint may be a Phi
node, meaning that a number of constraints arrives at that point
for the same nodecl.

This commit adds a new attribute to edges determining whether an
edge is a back edge: its target is a node that appears before in the flow.

This also adds support in PCFG to DOT to print constraints information.

See #1862