ReachDefs must not shallow_copy unless necessary
ExpressionEvolutionVisitor requires information about reaching definitions
to decide whether or not an expression is adjacent or not. For that,
we need to look for the node in the pcfg corresponding to the reaching
definition. This search uses the nodecl pointer comparison and, when
we do shallow_copy a nodecl, the associated tree is no longer the same,
so the comparison will return false when we expect it to return true.