Fix clearing PCFG visits in level
This method must take into account the cases where we exit the
context we are cleaning up: the creation of a task. Here, we still
want to clear the visits of the task, but we need to change the
context because the code of the task is not enclosed in the same
context as the creation of the task.