Propagate T/F constraints until node with stmts
When a node creates two branches (such a loop condition), two
different constraints are generated, one for each branch.
When propagating this constraint to the corresponding node,
we need to be sure the target node contains statements or
is a graph node, otherwise (entry|exit nodes, etc.) the constraint
will be lost.