For non language reference types sizeof/alignof should use the base type
Otherwise we get emit a sizeof/alignof of a pointer, which is not the
expected semantic of the sizeof/alignof of a reference type.

While this change is not extra robust, as there could be other
non-language reference types as subtypes of a compound type they should
not affect the result of the sizeof/alignof.