Check parameters when generating TDG conditions
There is no reaching definition at a function's entry point for the
parameters of the function. Then, when building the TDG edges conditions,
we need to check whether a variable comes from a parameter and does not
have a reaching definition or simply has a reaching definition (any other
possibility must come from a wrong code|analysis).

See #1908