TDG adds data flow size on dependencies info
Each dependency edge add a new field 'size' containing the size
of the data creating the dependency(ies) between two tasks.
Those edges with no dependency (barrier or taskwait synchronizations)
do not add this field so far. We may want to have this new field in
all cases with a value of 0 when no dependency occurs.

See #2055.