Check nodecl validity when splitting objects usage
When it is not possible to split the usage of a variable,
the function performing this operation returns a null nodecl.
Since this function can be called several times using the value returned
in the previous call, we need to check whether the return is valid or not,
and stop iterating when the first value returned is a null nodecl.

Fixes #2077