Register all conversions
This commit fixes #2076

This commit is huge because registering conversions uncovered lots of
bugs in many parts of the compiler. In particular, Fortran FE was
oblivious to conversions and so was the Fortran codegen.

This commit fixes #1798

Also note that as of this commit, const_value_t API allows to distinguish
NULL-terminated strings from non NULL terminated ones. This is important
since in Fortran, literal strings do not end in NULL while C and C++ do.

Since we are bold people and mix the two languages in the Nanos++
transformation, we need a way to remember that a NULL-ended string
literal is being used in Fortran. Note that #1798 called for a new
const-value literal kind, but this would break Fortran modules and
probably is not worth since 99% of the time the NULL itself is
uninteresting to the compiler and the API is able to tell its client
if the string literal ends with a NULL.