Extend reach defs map to keep the original stmt
Reaching definitions of pre/post-in/de-crements and
combined assignements cannot be matched with the node in the
PCFG that originated the reaching definition.
We have extended the map that relates the variables with their
reaching definition by adding the statement that generated that
definition (only when the definition is different than the statement).