Reach defs initializes all variables as Unknown
Note: When an object has sub-objects, only the object is
initialized to Unknown. This means that initializing a sub-object
does not cause all other sub-objects to be materialized as
Unknown, it is assumed.

We are renaming the Nodecl::Undefined to Nodecl::Unknown.

This commit also modifies the function that compares reaching
definitions maps: the key must be structurally equal whereas the
value must be pointer equal.

See #2086