Use nodecl const value in expr_reduction, if exists
Nodecls which children are constant values, already have their
constant value computed. When reducing expressions, the base case is
when the nodecl already has the constant value computed (there is
no need to check the children).

This commit also replaces the call to Nodecl::Utils::replace by the
method replace of the class Nodecl::NodeclBase in ExpressionReduction class.

Fixes #2094.