Add list of control structures to TDG JSON
This change simplifies and reduces the replicated information about
the control structures involved in the execution of a task.

Now, information about control strucures are in two places:
* A list of the control structures: All those attributes belonging
to the control structure (identifier, type and locus) are defined
within the list of control structures.
* The control structures associated to a given task: those attributes
that affect the instantiation of a given task (taken branch if select
control structure, condition that must fulfill to instantiate the task)
are associated to the control structure of the task.

See #2081