Split analysis interface from implementation
This commit separate the interface of the analysis (tl-analysis-interface)
from the implementation of the analysis queries (tl-analysis-queries).

This is useful since some queries require the use of template functions
that have to be implemented in the header file. Therefore, if
we have the implementation in a different file, this functions won't
be included and instanciated each time the user needs to use the
analysis.

This structure also makes easier the inheritance (VectorizationAnalysisInterface).
Now, we can extend the interface while the implementation remains
independent.