Fix ptr & rfnc use-def for unreachable code funcs
This commit fixes and homogenizes IPA Use-Def calculation for
functions which code is not reachable.

A new function is added in Use-Def to simplify some patters in the
arguments of a function call:
* &*v -> v
* *&v -> v
* &v[0] -> v

A fake symbol is added to the Analysis Checker phase so we are able
to specify when an analysis set (ue, kill, live, etc.) must be empty.
This symbol is called __ANALYSIS_NONE__.

See #2101