Use-Def allows same variable in both UE and UNDEF
We propagating usage from inner to outer nodes, we must allow
the same variable to be in both UE and UNDEF sets. Otherwise,
if the variable has been UE but it is afterwards UNDEF and we delete
the variable from the UE set, we are losing a lot of information.

This commit also fixes PCFG cleaning proces for post_sync nodes