Keep scalar code when it's possible
This commit introduces major changes in the
vectorization strategy to maintain scalar code
as much as possible.

A new visitor visits local symbols and decided if
the type of the symbol must be vectorized or not.
With the scalar or vector TL::Type, the vectorization
algorithm decides if a particular operation must be
vectorized or keep scalar.

It fixes #2106