TDG json defines separatedly control structures
In order to avoid repeating the same information for all nodes within
the same control structure, we split the inherent information to the
control structure from that information attached to the task execution.

Because of this change, we need to differentiate between:
* IfElse: There can only be two exit branches, the one that fulfills
the condition and the one that does not fulfill. In this case, the
control structure contains the condition of the If statement and each
node refering to the control structure contains a list of branch_ids
(1 means that the node exists when the condition fulfills and 2 means
the opposite).
* Switch: There are many conditions, one for each case. In this case,
the control structure contains the condition of the Switch statement
and each node refering to the control structure contains a list of
branch_conds, with the conditions of the cases leading to that node.

See #2081

This commit also fixes the order how the control structures are showed
(the same order as in the source code).