Refactor type deduction so it is more conformant
Consider the following valid case

  template <typename T>
  struct A { };

  template <typename T>
  struct B : A<T> { };

  void f(A<T>&);

  void g()
  {
      B<int> b;
      f(b);
  }

we have to deduce T <- int.

Until now we were deducing using the bases in the unification step. But this
is wrong. In the following example we would deduce several values for T,
where only T <- int** is valid.

  template <int N, typename T>
  struct B { };

  template <int N, typename T>
  struct B<N, T*> : B<N+1, T> { };

  template <typename T>
  struct D : B<0, T> { };

  template <int N, typename T>
  void f(B<N, T>&) { }

  void t()
  {
      D<int**> d;

      f<0>(d);
  }

The problem is that B<0, int*> and B<0, int> are not bases of D<int**> (only
B<0, int**>, B<1, int*> and B<2, int> are). So a simple check where we attempt
to substitute the parameters will lack a check that the deduced argument type
is a base of the argument type.