Analysis adds initial support for Linear clause
A new node SimdFunction has been added to the PCFG.
This node can be associated with clauses such as linear.
Other simd nodes, such as SimdFor, may be associated with this clause, as well.

New queries have been added to the Analysis and the VectorizationAnalysis
interfaces allowing consulting about this property.

See #2113