Initial support for uniform and linear clauses
Induction variables queries are being migrated to
linear queries since IVs are a subset of linear variables.

Uniform and linear argument/paraments are kept scalar.

Uniform and linear clauses in functions are not working yet.

We have a restriction that would require further implementation work:
If a linear or uniform variable in a loop (determined by the compiler)
is used as argument in a function call, the function code must have
been annotated with the corresponding linear and uniform clause on
that parameter. Otherwise, the compilation will fail because called
and callee will have arguments/paramenters with different types
(scalar/vector respectively)

See #2075
See #2113