PCFG wo context prints switch properly
In the new implementation of the PCFG printing (without context nodes)
we assumed that all Entry nodes may have only one child. This is not
true in the case of switch nodes because of the context node appearing
between the condition of the switch and all the possible cases.

See #2154