Replace implicit Control Structure in JSON tasks
So far we had a virtual Control Structure (CS), called Blank, that
represented the case when a task was not enclosed in any CS.
Nonetheless, this approach does not solve the problem of keeping
the values of the variables when a task is created.
Now, we have deleted the concept of Blank CS and instead we have
an Implicit CS that (immediately) encloses any task. This way,
we only assign variable values to the CSs and the cases where a task
is not contained in any CS are covered by the Implit CS as well.

See #2081