Overhaul instantiation process
Now all functions that can be instantiated have an emission_template
field. This applies even for nonmember template functions, where
the emission template is going to be themselves.

Also add a bit stating that a template-parameter-list is actually
instantiated. This is needed because C++ does not allow redundant
template-parameters for instantiated stuff. On the other hand we cannot
collapse the nest of template parameters, otherwise some instantiation
(e.g. those of default arguments) will fail.