Modify PCFG clause to store the whole nodecl
So far we were storing only the arguments of the clause.
This caused weird sequences for clauses with different kind of parameters.

This commit also fixes the situation caused by commit
8bba4865143698b40830381cb7e651d97e7afd0f, were the clause InAlloca was removed.
Not all references to this clause were removed from the analysis phase.
We are removing them now and also adding support for clause SharedAndAlloca.