First step towards removing decl context from type comparison
This was only required when comparing convoluted dependent typenames
as it may require examining uninstantiated templates. This was
integrated in the code of equivalent_types forcing us to pass
a decl_context_t everywhere.

This causes two problems: type equality is contextual. Not really,
only in equivalent_types_in_context. equivalent_types called
the former using the global context. So, 90% of the times
type equality was not actually contextual.

The second: given that type equivalence is crucial when verifying
types, it must be fast. Requiring a context when comparing two
types is awful because we cannot speed up the comparison by using
amortized techniques (like mf-sets or hashes).

So instead of forcing type equivalence to carry a context, we
will convert the types in the only place of the code where such
context was required. This happens in C++ in the following case:

template <typename T>
struct B { };

template <typename T>
struct A {
    typename B<T>::S foo();
    typedef typename B<T>::S Q;
};

template <typename T>
typename A<T>::Q A<T>::foo()
{
}

Note that 'typename A<T>::Q' and 'typename B<T>::S' in the declaration of 'foo'
name the same entity. But we only know this if if examine the uninstantiated
template 'A<T>'. We can do this because we are exactly in the class-scope of
A<T>.

It is this special 'examination' of uninstantiated templates of dependent
typename's that was being done in equivalent_types_in_context. Now this is not
done anymore and the C++ FE fixes the type prior the equivalence checking.