Propagate ranges from Constraint Graph to PCFG
This commit add the mechanism of retrieving information from
range analysis to the PCFG. Storing this information in the PCFG
makes it usable from any other analysis phase.

We also removes from Entry/Exit edges of the PCFG the ranges information,
and add it for Taskwait and Barrier nodes.

See #1862