Lint support for incoherent dependency clauses
This commit adds support for checking incoherencies in the
data-dependency clauses of a task. Note that we do not check if there
are missing clauses, but whether the existing clauses are incoherent.
We may want to add support also for advising clauses, but this is
part of the AUTO_DEPS process.

See #2167