Add analysis checking for UNEFINED/UNKNOWN RDs
Global variables have an UNDEFINED reaching definition
(unless they are defined within the function being analyzed),
because we cannot ensure these variables have a value at that point.
Parameters have an UNKNOWN reaching definition (unless they are defined
within the function being analyzed), because we are sure these variables
have a value, but we do not know it.

This commit also fixes the computation of the Reaching_Definitions_In set
of a graph node (instead of recaulculating it from the predecessors,
we get the value from joining the entry's children). Otherwise, we missed
UNKNOWN and UNDEFINED reaching definitions.

We also add a test for this feature and adapt the Analysis Check phase
to deal with UNKNOWN and UNDEFINED reaching definitions (we use fake
variables to specify these values in the assert clauses).

See #2107.