Correct next_sync for tasks seq. concurrent code
When Correctness analysis looks for sequential code concurrent
with a task, it traverses the different paths connecting all
"last_sync" and "next_sync" points found during tasks synchronization
analysis. When a "next_sync" point is a task, we will never reach it,
because when looking for sequential code we do not traverse tasks.
Thus, when we found any "next_sync" which is a task, we recursively
look for the points where this task synchronizes until we find points
where the sequential code will reach (taskwaits, barriers, virtual
synchronizations), and use them instead.

This is a temporary fix because it makes much more sense to implement
this in the tasks synchronization analysis, but that will cause many
other changes.

See #2167