Refactor type deduction
In the context of the new C++2011 standard, the deduction code of
Mercurium was starting to rot. The old code did not follow the standard,
but instead attempted to be "intuitive". Because of this slackness, some
codes were incorrectly accepted, some other were incorrectly rejected.

The new code, tries to follow in a closer way the latest C++ draft
(September 2014). Changes affect deduction and partial ordering of
classes and templates.

The old type/expression unification routines in cxx-typeunif.[ch] have
been removed. Instead all the deduction happens now in cxx-typededuc.c.

Also lots of efforts have been put in trying to save memory, given that
the old code leaked lots of data when deducing types. There are still
some leaks caused by the typesystem but we will address at a later
moment.

During the implementation, several inconsistencies in the Standard
arose. In particular g++, clang++ and icpc, differ in the interpretation
of template parameter packs. Mercurium implementation first tries to
follow the standard but in some edge cases we follow the consensus.