Minor improvement to -v compatibility with gcc
Only clear link data if there are no actual input files (no source).
This is for invocations like

  mcc -v -lfoo

where we have to strip -lfoo otherwise gcc attempts to link. Doing this
unconditionally is problematic when the user legitimally wants to see
what's going on, like in:

  mcc -v -o program program.c -lfoo

According to Florentino this patch seems enough.

This commit fixes #1983