Create an explicit context for all constructs
In Fortran, some control structures do not have new static contexts.
This is usually not a problem unless we start transforming the code
and we generated new variables in deeper contexts.

Now the Fortran FE will create new contexts in control structures
so their shape is closer to the shape of the control structures
created by the C and C++ contexts (which almost always include
a context).

This commit fixes #2177